+++ Changelog.md
@@ -1,5 +1,6 @@
 # NeuroScheme
 
 ## git master
+* [!5] Added scoop for color mappers
 * [!4] Enabled horizontal and vertical splitting of panes
 * First version of the refactor of legacy neuroscheme